Moved everything related to AssignReferenceFromCollector to e.e.cst.functions.collector
