+++ cst/plugins/eu.esdihumboldt.cst.functions.collector/plugin.xml
@@ -6,9 +6,37 @@
       <propertyTransformation
             class="eu.esdihumboldt.cst.functions.collector.AssignReferenceFromCollector"
             engine="eu.esdihumboldt.align.java"
-            function="eu.esdihumboldt.hale.common.align.collector.assign"
-            id="eu.esdihumboldt.hale.common.align.collector.assign">
+            function="eu.esdihumboldt.cst.functions.collector.assign"
+            id="eu.esdihumboldt.cst.functions.collector.assign">
       </propertyTransformation>
+   </extension>
+   <extension
+         point="eu.esdihumboldt.hale.align.function">
+      <propertyFunction
+            category="eu.esdihumboldt.hale.align.general"
+            icon="icons/augmentation.gif"
+            identifier="eu.esdihumboldt.cst.functions.collector.assign"
+            name="Assign collected references">
+         <targetProperties>
+            <property
+                  eager="false"
+                  maxOccurrence="1"
+                  minOccurrence="1">
+            </property>
+         </targetProperties>
+         <functionParameter
+               deprecated="false"
+               description="Name of the collector that contains the references"
+               label="Collector name"
+               maxOccurrence="1"
+               minOccurrence="1"
+               name="collector"
+               scriptable="false">
+            <parameterBinding
+                  class="java.lang.String">
+            </parameterBinding>
+         </functionParameter>
+      </propertyFunction>
    </extension>
 
 </plugin>
